Set up separate component for active shows dropdown
