+++ frontend/src/components/ClosedShowsDropdown.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="mt-4">
+  <div>
     <h3 class="text-[#393f4d] font-semibold text-center md:text-left">
       Closed Shows:
       <span class="text-blue-500 cursor-pointer hover:underline" @click="toggleDropdown">
@@ -68,7 +68,7 @@ const formatTourDate = (dateString) => {
 };
 </script>
   
-  <style scoped>
+<style scoped>
 .closed-shows-dropdown {
   position: absolute;
   z-index: 50;
